Android compatibility

Former-commit-id: 3601bdb524c6193f038fb6536f68fb205d90df8f
Former-commit-id: 122a081d0a00fc8042565782ffe94796330d2eda [formerly e7f96c4bf6a2677c7ad727e0c0d7e3ef54f11fa6]
Former-commit-id: edf1d1b669b754d4250cc19d29ce079ab44eb021

+++ modules/renderer/source/hip/hiprenderer/backend/d3d/d3drenderer.d
@@ -334,7 +334,7 @@ class Hip_D3D11_Renderer : IHipRendererImpl
     }
     public final bool isRowMajor(){return false;}
 
-    public bool hasErrorOccurred(out string err, string file = __FILE__, int line = __LINE__)
+    public bool hasErrorOccurred(out string err, string file = __FILE__, size_t line = __LINE__)
     {
         if(hasDebugAvailable)
         {
